+++ mobile/android/android-components/components/feature/top-sites/src/main/java/mozilla/components/feature/top/sites/TopSitesConfig.kt
@@ -11,11 +11,13 @@ import mozilla.components.concept.storage.FrecencyThresholdOption
  * whether or not to include top frecent sites in the top sites feature.
  *
  * @property totalSites A total number of sites that will be displayed.
+ * @property fetchProvidedTopSites Whether or not to fetch top sites from the [TopSitesProvider].
  * @property frecencyConfig If [frecencyConfig] is specified, only visited sites with a frecency
  * score above the given threshold will be returned. Otherwise, frecent top site results are
  * not included.
  */
 data class TopSitesConfig(
     val totalSites: Int,
+    val fetchProvidedTopSites: Boolean,
     val frecencyConfig: FrecencyThresholdOption?
 )