+++ mobile/android/android-components/components/feature/top-sites/src/main/java/mozilla/components/feature/top/sites/presenter/DefaultTopSitesPresenter.kt
@@ -46,6 +46,7 @@ internal class DefaultTopSitesPresenter(
         scope.launch {
             val topSites = storage.getTopSites(
                 innerConfig.totalSites,
+                innerConfig.fetchProvidedTopSites,
                 innerConfig.frecencyConfig
             )